internal/cryptsetup: provide ext4 setup
